[doku] Mention limitation in input validation for numbers. (TLE-14)
The input validation for numbers shown as text fields has a limitation: If the minimum value is not 0, you can enter any
value between 0 and the maximum value.

If the tag list contains an incorrect value, you may be unable to edit the value directly.
In this case, mark the value and press backspace, then enter a correct value.

Put the instructions for building an nxo file at the end of the text.