[code] Rename the tag RCX_TAG_SERVX_PORT_NUMBER to HIL_TAG_HTTP_PORT_CONFIG (TLE-84)
Rename the tag, data structure definition, the strings visible in the GUI and in tagtool output.
Add a check to the tagtool "edit" command to recognize patch files that have the old name in them.
When the old tag name is detected, change it to the new one and also change the name of the only
structure member to "usPort".